rvl: remove fabricated section-padding hacks; fix absorbed symbol sizes
Retail data slices end at the next unit's alignment boundary, so most
per-unit .sdata/.sbss/.bss/.data slices carry trailing zero pad bytes.
Those are linker-inserted inter-unit alignment, not content, and are
reproduced automatically at link time - the original sources never
contained them. Remove ~60 fabricated constructs across RVL_SDK:

- fake .init-section functions keeping "\0\0\0" strings in .sdata
  (GXInit, i2c, OSPlayRecord, encutility, OSFont, dsp, OSAlloc, OSIpc,
  EXIBios, ai)
- file-scope pad globals / forced-section zero objects / sdata_threshold
  pragma pads (AX*, bte, hbm/lyt, OS, dvd, exi, si, nand, ipc, sc, ...)
- trailing \0 padding inside string literals and padded array sizes
- struct/array pad fabrication (lowDone, __AXStudioBlock, Debug_BBA[8],
  axDspSlaveLength u32->u16, s_haltFmt[8], s_arcFileName[8], ...)

The old DOL was inflated by these (+0x60 .sdata, +0x100 .text); FULL_MATCH
acceptance is function-level and does not depend on the pad bytes. objdiff
per-unit: no code regressions (verified all touched units; several data%
improved).

Accounting fixes where ppcdis absorbed the pad into a symbol size:
- config/us/symbols.txt: correct 260 data:string sizes to first-NUL (regen
  build objects via dtk dol split; it skips existing files)
- tools/postprocess_reloc_names.py: NANDLogging lbl_805512D4 0x2C -> 0x27
  etc. (baked-in absorbed sizes)

Collateral: fix committed compile error in DSPCode.c (u32/u16
redeclaration), SbssPad symbol collision (scsystem/dvd_broadway), define
AXFXChorusExpDpl2's own .sdata2 float pool (externs were never defined).
Docs: MWCC_REFERENCE section on why section padding must never be
fabricated in source.

+++ libs/RVL_SDK/src/revolution/ax/DSPCode.c
@@ -518,4 +518,4 @@ u8 axDspSlave[] = {
 
 u16 axDspInitVector = 0x0010;
 u16 axDspResumeVector = 0x0037;
-u32 axDspSlaveLength = ARRAY_SIZE(axDspSlave); /* u32: low u16 + 2 zero pad = retail .sdata 0x6 -> 0x8 */
+u16 axDspSlaveLength = ARRAY_SIZE(axDspSlave);
